# Create ReadAccessor benchmark