--- a/benchmark/CMakeLists.txt
+++ b/benchmark/CMakeLists.txt
@@ -196,6 +196,7 @@ set(SWIFT_BENCH_MODULES
     single-source/WordCount
     single-source/XorLoop
     cxx-source/CreateObjects
+    cxx-source/ReadAccessor
 )
 
 set(SWIFT_MULTISOURCE_SWIFT_BENCHES