--- a/benchmark/utils/CxxTests/Subscripts.h
+++ b/benchmark/utils/CxxTests/Subscripts.h
@@ -1,0 +1,10 @@
+#ifndef BENCHMARK_SUBSCRIPTS_H
+#define BENCHMARK_SUBSCRIPTS_H
+#include <vector>
+
+using TwoDimensionalVector = std::vector<std::vector<int>>;
+
+inline TwoDimensionalVector initVector() { return {100, std::vector<int>{1000, 0}}; }
+
+#endif
+