--- a/benchmark/utils/CxxTests/module.modulemap
+++ b/benchmark/utils/CxxTests/module.modulemap
@@ -2,3 +2,8 @@ module CxxCreateObjects {
   header "CreateObjects.h"
   requires cplusplus
 }
+
+module CxxSubscripts {
+  header "Subscripts.h"
+  requires cplusplus
+}