Create a benchmark to test the
performance of synthesized read
accessors generated to interop
with C++ [] operators.