@NuriAmari 2022-03-07
I will look into it, but with the previous large non trivial struct, the benchmark did show performance improvements, but with this vector approach I'm not seeing anything.