@zoecarver 2022-03-07
> I will look into it, but with the previous large non trivial struct, the benchmark did show performance improvements, but with this vector approach I'm not seeing anything.

Hmm that's really odd. Let's see what the CI says, but vector is really what we want to optimize. I guess we can look into the IR and see what's going on. If there isn't an improvement, I guess we can add both benchmarks 🤷‍♂️