@NuriAmari 2022-03-07
I am looking at the IR, the read accessor is generated properly, but for some reason the caller decides to copy from the address anyways.

**Read Accessor:**

```
// std.__1.__CxxTemplateInstNSt3__16vectorI10NonTrivialNS_9allocatorIS1_EEEE.subscript.read
sil shared [transparent] @$sSo3stdO3__1O65__CxxTemplateInstNSt3__16vectorI10NonTrivialNS_9allocatorIS1_EEEEVySo0hI0VSicir : $@yield_once @convention(method) (Int, @in_guaranteed std.__1.__CxxTemplateInstNSt3__16vectorI10NonTrivialNS_9allocatorIS1_EEEE) -> @yields @in_guaranteed NonTrivial {
// %0 "__n"                                       // users: %5, %2
// %1 "self"                                      // users: %5, %3
bb0(%0 : $Int, %1 : $*std.__1.__CxxTemplateInstNSt3__16vectorI10NonTrivialNS_9allocatorIS1_EEEE):
  debug_value %0 : $Int, let, name "__n", argno 1 // id: %2
  debug_value %1 : $*std.__1.__CxxTemplateInstNSt3__16vectorI10NonTrivialNS_9allocatorIS1_EEEE, let, name "self", argno 2, implicit, expr op_deref // id: %3
  // function_ref _ZNKSt3__16vectorI10NonTrivialNS_9allocatorIS1_EEEixEm
  %4 = function_ref @_ZNKSt3__16vectorI10NonTrivialNS_9allocatorIS1_EEEixEm : $@convention(cxx_method) (Int, @in_guaranteed std.__1.__CxxTemplateInstNSt3__16vectorI10NonTrivialNS_9allocatorIS1_EEEE) -> UnsafePointer<NonTrivial> // user: %5
  %5 = apply %4(%0, %1) : $@convention(cxx_method) (Int, @in_guaranteed std.__1.__CxxTemplateInstNSt3__16vectorI10NonTrivialNS_9allocatorIS1_EEEE) -> UnsafePointer<NonTrivial> // user: %6
  %6 = struct_extract %5 : $UnsafePointer<NonTrivial>, #UnsafePointer._rawValue // user: %7
  %7 = pointer_to_address %6 : $Builtin.RawPointer to [strict] $*NonTrivial // user: %8
  %8 = begin_access [read] [unsafe] %7 : $*NonTrivial // users: %10, %13, %9
  yield %8 : $*NonTrivial, resume bb1, unwind bb2 // id: %9
```

**Usage:**

```
  // function_ref _ZNKSt3__16vectorI10NonTrivialNS_9allocatorIS1_EEEixEm
  %38 = function_ref @_ZNKSt3__16vectorI10NonTrivialNS_9allocatorIS1_EEEixEm : $@convention(cxx_method) (Int, @in_guaranteed std.__1.__CxxTemplateInstNSt3__16vectorI10NonTrivialNS_9allocatorIS1_EEEE) -> UnsafePointer<NonTrivial> // user: %39
  %39 = apply %38(%35, %36) : $@convention(cxx_method) (Int, @in_guaranteed std.__1.__CxxTemplateInstNSt3__16vectorI10NonTrivialNS_9allocatorIS1_EEEE) -> UnsafePointer<NonTrivial> // user: %40
  %40 = struct_extract %39 : $UnsafePointer<NonTrivial>, #UnsafePointer._rawValue // user: %41
  %41 = pointer_to_address %40 : $Builtin.RawPointer to [strict] $*NonTrivial // user: %42
  %42 = begin_access [read] [unsafe] %41 : $*NonTrivial // users: %44, %45
  %43 = alloc_stack $NonTrivial                   // users: %59, %50, %48, %44
  copy_addr %42 to [initialization] %43 : $*NonTrivial // id: %44
  end_access %42 : $*NonTrivial                   // id: %45
```

No idea why yet, but I'll keep looking. Maybe something specific to `std::vector`?