@zoecarver 2022-03-07
From the CI:

### Performance (x86_64): -O

**Regression**       | **OLD** | **NEW** | **DELTA** | **RATIO**
:---                 | ---:   | ---:   | ---:   | ---:     
Dictionary4          | 152    | 193    | +27.0% | **0.79x (?)**
Dictionary4OfObjects | 206    | 239    | +16.0% | **0.86x (?)**
Calculator           | 109    | 122    | +11.9% | **0.89x (?)**
&nbsp; | | | | 
**Improvement**      | **OLD** | **NEW** | **DELTA** | **RATIO**
FlattenListFlatMap   | 3513   | 3201   | -8.9%  | **1.10x (?)**
&nbsp; | | | | 
**Added**            | **MIN** | **MAX** | **MEAN** | **MAX_RSS**
ReadAccessor         | 117031 | 117396 | 117204 | —      

### Code size: -O


### Performance (x86_64): -Osize

**Regression**    | **OLD** | **NEW** | **DELTA** | **RATIO**
:---              | ---:   | ---:   | ---:   | ---:     
Array2D           | 4288   | 4704   | +9.7%  | **0.91x (?)**
&nbsp; | | | | 
**Improvement**   | **OLD** | **NEW** | **DELTA** | **RATIO**
DropWhileSequence | 15     | 14     | -6.7%  | **1.07x (?)**
SetUnionInt25     | 76     | 71     | -6.6%  | **1.07x (?)**
&nbsp; | | | | 
**Added**         | **MIN** | **MAX** | **MEAN** | **MAX_RSS**
ReadAccessor      | 122013 | 125432 | 123553 | —      

### Code size: -Osize


### Performance (x86_64): -Onone

**Improvement**        | **OLD** | **NEW** | **DELTA** | **RATIO**
:---                   | ---:   | ---:   | ---:   | ---:     
ParseFloat.Float.Exp   | 11     | 10     | -9.1%  | **1.10x (?)**
DataReplaceSmallBuffer | 3000   | 2800   | -6.7%  | **1.07x (?)**
&nbsp; | | | | 
**Added**              | **MIN** | **MAX** | **MEAN** | **MAX_RSS**
ReadAccessor           | 445319 | 452412 | 448067 | —      

### Code size: -swiftlibs



✅  | Benchmark Check Report
---|---
⛔️⏱ | `ReadAccessor` has setup overhead of 59346 μs (51.0%). <br><sub> Move initialization of benchmark data to the `setUpFunction` registered in `BenchmarkInfo`.
⛔️⏱ | `ReadAccessor` execution took at least 57045 μs (excluding the setup overhead). <br><sub> Decrease the workload of `ReadAccessor` by a factor of 64 (100), to be less than 1000 μs.
⚠️Ⓜ️ | `ReadAccessor` has very wide range of memory used between independent, repeated measurements. <br><sub> `ReadAccessor` mem_pages [i1, i2]: min=[6, 6] 𝚫=0 R=[19, 9]


<details>
  <summary><strong>How to read the data</strong></summary>
The tables contain differences in performance which are larger than 8% and
differences in code size which are larger than 1%.

If you see any unexpected regressions, you should consider fixing the
regressions before you merge the PR.

**Noise**: Sometimes the performance results (not code size!) contain false
alarms. Unexpected regressions which are marked with '(?)' are probably noise.
If you see regressions which you cannot explain you can try to run the
benchmarks again. If regressions still show up, please consult with the
performance team (@eeckstein).
</details>

<details>
  <summary><strong>Hardware Overview</strong></summary>

      Model Name: Mac mini
      Model Identifier: Macmini8,1
      Processor Name: 6-Core Intel Core i7
      Processor Speed: 3.2 GHz
      Number of Processors: 1
      Total Number of Cores: 6
      L2 Cache (per Core): 256 KB
      L3 Cache: 12 MB
      Memory: 32 GB

</details>