@NuriAmari 2022-03-07
> @NuriAmari maybe try addressing those warning at the bottom as well. Not sure exactly why the copy is there. Is that with optimizations enabled?

Not sure what to do about the last one, but the first two I think I can fix.