@zoecarver 2022-03-07
> Not sure what to do about the last one, but the first two I think I can fix.

Yeah me neither. I think it's fine to leave it. I suspect making the vector a bit smaller would help, but other than that, idk. Maybe 1000 elements instead?