@zoecarver 2022-03-07
### Performance (x86_64): -O

**Regression**                | **OLD** | **NEW** | **DELTA** | **RATIO**
:---                          | ---:  | ---:  | ---:   | ---:     
FlattenListLoop               | 1004  | 1580  | +57.4% | **0.64x (?)**
Dictionary4                   | 153   | 192   | +25.5% | **0.80x**
FlattenListFlatMap            | 3612  | 4332  | +19.9% | **0.83x (?)**
Dictionary4OfObjects          | 207   | 239   | +15.5% | **0.87x (?)**
Calculator                    | 109   | 122   | +11.9% | **0.89x (?)**
DropWhileAnySequence          | 1251  | 1354  | +8.2%  | **0.92x (?)**
&nbsp; | | | | 
**Improvement**               | **OLD** | **NEW** | **DELTA** | **RATIO**
IterateData                   | 872   | 807   | -7.5%  | **1.08x (?)**
Set.subtracting.Seq.Empty.Int | 122   | 114   | -6.6%  | **1.07x (?)**
&nbsp; | | | | 
**Added**                     | **MIN** | **MAX** | **MEAN** | **MAX_RSS**
ReadAccessor                  | 11527 | 11772 | 11612  | —      

### Code size: -O


### Performance (x86_64): -Osize

**Regression**                 | **OLD** | **NEW** | **DELTA** | **RATIO**
:---                           | ---:  | ---:  | ---:   | ---:     
Array2D                        | 4304  | 4688  | +8.9%  | **0.92x (?)**
PrefixWhileAnySequence         | 186   | 201   | +8.1%  | **0.93x (?)**
&nbsp; | | | | 
**Improvement**                | **OLD** | **NEW** | **DELTA** | **RATIO**
FlattenListLoop                | 1416  | 1028  | -27.4% | **1.38x (?)**
FlattenListFlatMap             | 3119  | 2717  | -12.9% | **1.15x (?)**
StringBuilderWithLongSubstring | 1070  | 990   | -7.5%  | **1.08x (?)**
&nbsp; | | | | 
**Added**                      | **MIN** | **MAX** | **MEAN** | **MAX_RSS**
ReadAccessor                   | 11743 | 12184 | 11890  | —      

### Code size: -Osize


### Performance (x86_64): -Onone

**Added**    | **MIN** | **MAX** | **MEAN** | **MAX_RSS**
:---         | ---:  | ---:  | ---:  | ---:   
ReadAccessor | 50167 | 50491 | 50298 | —    

### Code size: -swiftlibs



✅  | Benchmark Check Report
---|---
⛔️⏱ | `ReadAccessor` has setup overhead of 5850 μs (51.1%). <br><sub> Move initialization of benchmark data to the `setUpFunction` registered in `BenchmarkInfo`.
⚠️⏱ | `ReadAccessor` execution took at least 5588 μs (excluding the setup overhead). <br><sub> Decrease the workload of `ReadAccessor` by a factor of 8 (10), to be less than 1000 μs.


<details>
  <summary><strong>How to read the data</strong></summary>
The tables contain differences in performance which are larger than 8% and
differences in code size which are larger than 1%.

If you see any unexpected regressions, you should consider fixing the
regressions before you merge the PR.

**Noise**: Sometimes the performance results (not code size!) contain false
alarms. Unexpected regressions which are marked with '(?)' are probably noise.
If you see regressions which you cannot explain you can try to run the
benchmarks again. If regressions still show up, please consult with the
performance team (@eeckstein).
</details>

<details>
  <summary><strong>Hardware Overview</strong></summary>

      Model Name: Mac mini
      Model Identifier: Macmini8,1
      Processor Name: 6-Core Intel Core i7
      Processor Speed: 3.2 GHz
      Number of Processors: 1
      Total Number of Cores: 6
      L2 Cache (per Core): 256 KB
      L3 Cache: 12 MB
      Memory: 32 GB

</details>